Fixed warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@55315 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objects/seqalign/Dense_diag.hpp
@@ -125,7 +125,7 @@ inline
 CDense_diag::TDim CDense_diag::CheckNumRows() const
 {
     TDim dim = GetDim();
-    if (dim != GetIds().size()  ||  dim != GetStarts().size()) {
+    if (size_t(dim) != GetIds().size() || size_t(dim) != GetStarts().size()) {
         NCBI_THROW(CSeqalignException, eInvalidAlignment,
                    "CDense_diag::CheckNumRows()"
                    " dim is not consistent with ids.size & starts.size");
